refactor: Change Giphy file structure and update html tags
